s3ext: add a standalone config checking app
Usage: s3chkcfg -c "s3://endpoint/bucket/prefix config=path_to_config_file", to check the configuration.
       s3chkcfg -d "s3://endpoint/bucket/prefix config=path_to_config_file", to download and output to stdout.
       s3chkcfg -t, to show the config template.
       s3chkcfg -h, to show this help.

Also did refactoring to reuse functions common with s3ext module.

+++ gpAux/extensions/gps3ext/src/s3downloader.h
@@ -112,7 +112,7 @@ class Downloader {
     uint64_t readlen;
 
     unsigned char magic_bytes[4];
-    uint8_t magic_bytes_num;
+    uint64_t magic_bytes_num;
     compression_type_t compression;
     bool set_compression();
